Replace int with size_t for length

+++ aead.h
@@ -10,15 +10,15 @@ typedef struct {
 }state_t;
 
 size_t crypto_aead_encrypt(unsigned char* c, unsigned long long* clen,
-                           const unsigned char* m, unsigned long long mlen,
-                           const unsigned char* ad, unsigned long long adlen,
+                           const unsigned char* m, size_t mlen,
+                           const unsigned char* ad, size_t adlen,
                            const unsigned char* nsec,
                            const unsigned char*npub,
                            const unsigned char* k);
 
 size_t crypto_aead_decrypt(unsigned char* m, unsigned long long *mlen,
-                           const unsigned char* c, unsigned long long clen,
-                           const unsigned char* ad, unsigned long long adlen,
+                           const unsigned char* c, size_t clen,
+                           const unsigned char* ad, size_t adlen,
                            unsigned char* nsec,
                            const unsigned char* npub,
                            const unsigned char* k);
